Split scripts and dont use tx
